new module fetch added
